Reorganize symbols
