+++ packages/seroval/src/constants.ts
@@ -1,3 +1,52 @@
+export const enum SerovalConstant {
+  Null = 0,
+  Undefined = 1,
+  True = 2,
+  False = 3,
+  NegativeZero = 4,
+  Infinity = 5,
+  NegativeInfinity = 6,
+  NaN = 7,
+}
+
+export const enum SerovalNodeType {
+  Number = 0,
+  String = 1,
+  Constant = 2,
+  BigInt = 3,
+  IndexedValue = 4,
+  Date = 5,
+  RegExp = 6,
+  Set = 7,
+  Map = 8,
+  Array = 9,
+  Object = 10,
+  NullConstructor = 11,
+  Promise = 12,
+  Error = 13,
+  AggregateError = 14,
+  TypedArray = 15,
+  BigIntTypedArray = 16,
+  WKSymbol = 17,
+  URL = 18,
+  URLSearchParams = 19,
+  Reference = 20,
+  ArrayBuffer = 21,
+  DataView = 22,
+  Blob = 23,
+  File = 24,
+  Headers = 25,
+  FormData = 26,
+  Boxed = 27,
+}
+
+export const enum SerovalObjectFlags {
+  None = 0,
+  NonExtensible = 1,
+  Sealed = 2,
+  Frozen = 3,
+}
+
 export const enum Symbols {
   AsyncIterator = 0,
   HasInstance = 1,
