+++ packages/seroval/src/tree/deserialize.ts
@@ -8,7 +8,7 @@ import type { BigIntTypedArrayValue, TypedArrayValue } from '../types';
 import { deserializeConstant } from './constants';
 import { getReference } from './reference';
 import { getErrorConstructor, getTypedArrayConstructor } from './shared';
-import { SYMBOL_REF } from './symbols';
+import { SYMBOL_REF } from '../symbols';
 import type {
   SerovalAggregateErrorNode,
   SerovalArrayBufferNode,
