+++ packages/seroval/src/tree/sync.ts
@@ -53,10 +53,10 @@ import type {
   SerovalSetNode,
 } from './types';
 import {
-  SerovalNodeType,
   SerovalObjectRecordSpecialKey,
 } from './types';
 import { createURLNode, createURLSearchParamsNode } from './web-api';
+import { SerovalNodeType } from '../constants';
 
 type ObjectLikeNode = SerovalObjectNode | SerovalNullConstructorNode;
