+++ packages/seroval/src/types.ts
@@ -1,4 +1,4 @@
-import type { WellKnownSymbols } from './tree/symbols';
+import type { WellKnownSymbols } from './symbols';
 
 // Values that are non-recursive
 export type PrimitiveValue =
